+++ CHANGELOG.md
@@ -42,6 +42,7 @@ All notable changes to this project will be documented in this file.
     - sspmod_perun_IdpListsServiceDB to SimpleSAML\Module\perun\IdpListsServiceDB
     - sspmod_perun_LdapConnector to SimpleSAML\Module\perun\LdapConnector
     - sspmod_perun_RpcConnector to SimpleSAML\Module\perun\RpcConnector
+ - Added disco-tpl template file
  
  [Fixed]
  - Fixed the bug generating Array to string conversion Exception in PerunAttributes,